Modefy to get all content and change to V2 (#8724)
