+++ lib/radar-rules.js
@@ -2799,31 +2799,6 @@ module.exports = {
             },
         ],
     },
-    'cqwu.net': {
-        _name: '重庆文理学院',
-        www: [
-            {
-                title: '通知',
-                docs: 'https://docs.rsshub.app/university.html#chong-qing-wen-li-xue-yuan',
-                source: '/:type',
-                target: (params) => {
-                    if (params.type === 'channel_7721.html') {
-                        return '/cqwu/news/notify';
-                    }
-                },
-            },
-            {
-                title: '学术活动',
-                docs: 'https://docs.rsshub.app/university.html#chong-qing-wen-li-xue-yuan',
-                source: '/:type',
-                target: (params) => {
-                    if (params.type === 'channel_7722.html') {
-                        return '/cqwu/news/academiceve';
-                    }
-                },
-            },
-        ],
-    },
     'trakt.tv': {
         _name: 'Trakt.tv',
         '.': [
