+++ lib/router.js
@@ -720,9 +720,6 @@ router.get('/cqu/youth/:category', lazyloadRouteHandler('./routes/universities/c
 router.get('/cqu/sci/:category', lazyloadRouteHandler('./routes/universities/cqu/sci/info'));
 router.get('/cqu/net/:category', lazyloadRouteHandler('./routes/universities/cqu/net/info'));
 
-// 重庆文理学院
-router.get('/cqwu/news/:type?', lazyloadRouteHandler('./routes/universities/cqwu/news'));
-
 // 南京信息工程大学
 router.get('/nuist/bulletin/:category?', lazyloadRouteHandler('./routes/universities/nuist/bulletin'));
 router.get('/nuist/jwc/:category?', lazyloadRouteHandler('./routes/universities/nuist/jwc'));
