+++ lib/v2/cqwu/maintainer.js
@@ -1,0 +1,3 @@
+module.exports = {
+    '/news/:type?': ['Fatpandac'],
+};
