+++ lib/v2/cqwu/radar.js
@@ -1,0 +1,27 @@
+module.exports = {
+    'cqwu.net': {
+        _name: '重庆文理学院',
+        www: [
+            {
+                title: '通知',
+                docs: 'https://docs.rsshub.app/university.html#chong-qing-wen-li-xue-yuan',
+                source: '/:type',
+                target: (params) => {
+                    if (params.type === 'channel_7721.html') {
+                        return '/cqwu/news/notify';
+                    }
+                },
+            },
+            {
+                title: '学术活动',
+                docs: 'https://docs.rsshub.app/university.html#chong-qing-wen-li-xue-yuan',
+                source: '/:type',
+                target: (params) => {
+                    if (params.type === 'channel_7722.html') {
+                        return '/cqwu/news/academiceve';
+                    }
+                },
+            },
+        ],
+    },
+};
