+++ lib/v2/cqwu/router.js
@@ -1,0 +1,3 @@
+module.exports = function (router) {
+    router.get('/news/:type?', require('./index'));
+};
